requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
# Google GenAI SDK (v1beta)
google-genai
# Computer Vision & Audio
opencv-python
pyaudio
sounddevice
pillow
mss
# OCR (PaddleOCR, local)
paddlepaddle
paddleocr
taskgroup
exceptiongroup
# Browser Automation
playwright
# Smart Home
python-kasa
# Utilities
python-dotenv
# Face & Hand tracking
mediapipe
